+++ protostar/commands/test/test_command.py
@@ -8,12 +8,11 @@ from protostar.commands.test.test_collector_summary_formatter import (
     format_test_collector_summary,
 )
 from protostar.commands.test.test_result_formatter import format_test_result
-from protostar.commands.test.test_runner import TestRunner
 from protostar.commands.test.test_scheduler import TestScheduler
 from protostar.commands.test.testing_live_logger import TestingLiveLogger
 from protostar.commands.test.testing_seed import determine_testing_seed
 from protostar.compiler import ProjectCairoPathBuilder
-from protostar.test_runner import TestCollector, TestingSummary, TestResult
+from protostar.test_runner import TestCollector, TestingSummary, TestResult, TestRunner
 from protostar.utils.compiler.pass_managers import (
     StarknetPassManagerFactory,
     TestCollectorPassManagerFactory,
