+++ protostar/commands/test/test_scheduler.py
@@ -2,10 +2,10 @@ import multiprocessing
 import signal
 from typing import TYPE_CHECKING, Callable, List
 
-from protostar.commands.test.test_runner import TestRunner
 from protostar.commands.test.test_shared_tests_state import SharedTestsState
 from protostar.commands.test.testing_live_logger import TestingLiveLogger
 from protostar.commands.test.testing_seed import Seed
+from protostar.test_runner import TestRunner
 
 if TYPE_CHECKING:
     from protostar.test_runner import TestCollector
